+++ examples/micro-math-service/src/modules/math/index.js
@@ -2,26 +2,25 @@ function mathService(base) {
 
   const add = {
     name: 'sum',
-    path: '',
     handler: (msg, reply) => {
-      var sum = msg.left + msg.right;
-      return reply({ answer: sum, host: process.env.HOSTNAME });
+      var result = msg.left + msg.right;
+      return reply(base.utils.genericResponse({ answer: result, host: process.env.HOSTNAME }));
     }
   };
 
   const multiply = {
     name: 'multiply',
     handler: (msg, reply) => {
-      var sum = msg.left * msg.right;
-      return reply({ answer: sum, host: process.env.HOSTNAME });
+      var result = msg.left * msg.right;
+      return reply(base.utils.genericResponse({ answer: result, host: process.env.HOSTNAME }));
     }
   };
 
   const divide = {
     name: 'divide',
-    handler: (msg, reply) => {
-      var sum = msg.left / msg.right;
-      return reply({ answer: sum, host: process.env.HOSTNAME });
+    handler: (result, reply) => {
+      var result = msg.left / msg.right;
+      return reply(base.utils.genericResponse({ answer: result, host: process.env.HOSTNAME }));
     }
   };